ログイン状態に応じて LoginPage、IndexPage に遷移する
